Add Substitution Translations, fix frontend display and layout

+++ frontend/src/components/event/competition/excecution/Substitutions.tsx
@@ -32,6 +32,8 @@ import SwapHorizIcon from '@mui/icons-material/SwapHoriz'
 import {FormInputText} from '@components/form/input/FormInputText.tsx'
 import Add from '@mui/icons-material/Add'
 import DeleteIcon from '@mui/icons-material/Delete'
+import {useConfirmation} from '@contexts/confirmation/ConfirmationContext.ts'
+import FormInputLabel from '@components/form/input/FormInputLabel.tsx'
 
 type SubstitutionWithSwap = {
     substitution: SubstitutionDto
@@ -56,6 +58,8 @@ const Substitutions = ({reloadRoundDto, roundDto, roundIndex}: Props) => {
 
     const {eventId} = eventRoute.useParams()
     const {competitionId} = competitionRoute.useParams()
+
+    const {confirmAction} = useConfirmation()
 
     const substitutions: Array<SubstitutionWithSwap> = roundDto.substitutions
         .sort((a, b) => b.orderForRound - a.orderForRound)
@@ -125,10 +129,10 @@ const Substitutions = ({reloadRoundDto, roundDto, roundIndex}: Props) => {
         setSubmitting(false)
 
         if (error) {
-            feedback.error('todo ERROR')
+            feedback.error(t('event.competition.execution.substitution.add.error'))
         } else {
             closeDialog()
-            feedback.success('todo Saved')
+            feedback.success(t('event.competition.execution.substitution.add.success'))
         }
         reloadRoundDto()
     }
@@ -146,21 +150,29 @@ const Substitutions = ({reloadRoundDto, roundDto, roundIndex}: Props) => {
     }
 
     const handleDeleteSubstitution = async () => {
-        setSubmitting(true)
-        const {error} = await deleteSubstitution({
-            path: {
-                eventId: eventId,
-                competitionId: competitionId,
-            },
-        })
-        setSubmitting(false)
+        confirmAction(
+            async () => {
+                setSubmitting(true)
+                const {error} = await deleteSubstitution({
+                    path: {
+                        eventId: eventId,
+                        competitionId: competitionId,
+                    },
+                })
+                setSubmitting(false)
 
-        if (error) {
-            feedback.error('todo ERROR')
-        } else {
-            feedback.success('todo Deleted')
-        }
-        reloadRoundDto()
+                if (error) {
+                    feedback.error(t('event.competition.execution.substitution.delete.error'))
+                } else {
+                    feedback.success(t('event.competition.execution.substitution.delete.success'))
+                }
+                reloadRoundDto()
+            },
+            {
+                content: t('event.competition.execution.substitution.delete.confirmation'),
+                okText: t('common.delete'),
+            },
+        )
     }
 
     return (
@@ -168,7 +180,7 @@ const Substitutions = ({reloadRoundDto, roundDto, roundIndex}: Props) => {
             {roundIndex === 0 && (
                 <Box sx={{my: 2, flex: 1, display: 'flex', justifyContent: 'end'}}>
                     <Button variant={'outlined'} onClick={openDialog} startIcon={<Add />}>
-                        Add Substitution
+                        {t('event.competition.execution.substitution.add.add')}
                     </Button>
                 </Box>
             )}
@@ -211,7 +223,9 @@ const Substitutions = ({reloadRoundDto, roundDto, roundIndex}: Props) => {
                                 ) : (
                                     <Box sx={{display: 'flex', gap: 1, px: 2}}>
                                         <SouthIcon />
-                                        <Typography>for</Typography>
+                                        <Typography>
+                                            {t('event.competition.execution.substitution.inFor')}
+                                        </Typography>
                                         <NorthIcon />
                                     </Box>
                                 )}
@@ -233,14 +247,16 @@ const Substitutions = ({reloadRoundDto, roundDto, roundIndex}: Props) => {
                             <Box
                                 sx={{
                                     display: 'flex',
-                                    flexDirection: 'row',
                                     alignItems: 'center',
                                     justifyContent: 'space-between',
                                     [theme.breakpoints.up('lg')]: {
                                         width: 250,
                                     },
+                                    [theme.breakpoints.down('lg')]: {
+                                        flexDirection: 'column',
+                                    },
                                 }}>
-                                <Typography sx={{flex: 1}}>{sub.substitution.reason}</Typography>
+                                <Typography sx={{flex: 1, wordBreak: 'break-word'}} >{sub.substitution.reason}</Typography>
                                 {subIdx === 0 && roundIndex === 0 && (
                                     <IconButton sx={{ml: 2}} onClick={handleDeleteSubstitution}>
                                         <DeleteIcon />
@@ -254,7 +270,7 @@ const Substitutions = ({reloadRoundDto, roundDto, roundIndex}: Props) => {
             </List>
 
             <BaseDialog open={dialogOpen} onClose={closeDialog} maxWidth={'sm'}>
-                <DialogTitle>{'TODO Add Substitution'}</DialogTitle>
+                <DialogTitle>{t('event.competition.execution.substitution.add.add')}</DialogTitle>
                 <FormContainer formContext={formContext} onSuccess={onSubmit}>
                     <DialogContent dividers>
                         <Controller
@@ -269,24 +285,34 @@ const Substitutions = ({reloadRoundDto, roundDto, roundIndex}: Props) => {
                                 },
                             }) => (
                                 <Stack spacing={4}>
-                                    <Select
-                                        value={participantOutValue}
-                                        onChange={e => {
-                                            participantOutOnChange(e)
-                                        }}>
-                                        {subOutOptions.map(opt => (
-                                            <MenuItem key={opt?.id} value={opt?.id}>
-                                                {opt?.label}
-                                            </MenuItem>
-                                        ))}
-                                    </Select>
+                                    <FormInputLabel
+                                        label={t(
+                                            'event.competition.execution.substitution.participant',
+                                        )}
+                                        required={true}>
+                                        <Select
+                                            value={participantOutValue}
+                                            onChange={e => {
+                                                participantOutOnChange(e)
+                                            }}
+                                            sx={{width: 1}}>
+                                            {subOutOptions.map(opt => (
+                                                <MenuItem key={opt?.id} value={opt?.id}>
+                                                    {opt?.label}
+                                                </MenuItem>
+                                            ))}
+                                        </Select>
+                                    </FormInputLabel>
                                     {participantOutValue && (
                                         <SubstitutionSelectParticipantIn
                                             setupRoundId={roundDto.setupRoundId}
                                             selectedParticipantOut={participantOutValue}
                                         />
                                     )}
-                                    <FormInputText name={'reason'} label={'[todo] Reason'} />
+                                    <FormInputText
+                                        name={'reason'}
+                                        label={t('event.competition.execution.substitution.reason')}
+                                    />
                                 </Stack>
                             )}
                         />
